Adjusting post_save User model import
Moved post_save signal to __init__.
Added noqa flake8 to ignore unused imports.
